I did some changes to be more platform independent, e.g. introduced PRIu64
git-svn-id: file:///Volumes/Daten/wwwtech/cforum/trunk@1480 31cb981c-0fcd-0310-a313-f9fa081e70fd

+++ src/fo_view.c
@@ -37,6 +37,8 @@
 #include <sys/shm.h>
 #endif
 
+#include <inttypes.h>
+
 #include "readline.h"
 #include "hashlib.h"
 #include "utils.h"
@@ -242,9 +244,9 @@ void show_posting(cf_hash_t *head,void *shm_ptr,u_int64_t tid,u_int64_t mid)
   cf_set_variable(&tpl,cs,"postscript",ps->values[0],strlen(ps->values[0]),1);
   cf_set_variable(&tpl,cs,"regscript",reg->values[0],strlen(reg->values[0]),1);
 
-  len = snprintf(buff,256,"%llu",thread.tid);
+  len = snprintf(buff,256,"%" PRIu64,thread.tid);
   cf_set_variable(&tpl,cs,"tid",buff,len,0);
-  len = snprintf(buff,256,"%llu",thread.threadmsg->mid);
+  len = snprintf(buff,256,"%" PRIu64,thread.threadmsg->mid);
   cf_set_variable(&tpl,cs,"mid",buff,len,0);
 
   /* user values */
@@ -286,7 +288,7 @@ void show_threadlist(void *shm_ptr,cf_hash_t *head)
 #endif
 {
   /* {{{ variables */
-  int len;
+  size_t len;
   #ifndef CF_SHARED_MEM
   rline_t tsd;
   u_char *line,buff[128];
